removing even more junk code and updating tests
